# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tugas UKK Peminjaman Alat",
        "description": "Dokumentasi API Sistem Peminjaman Alat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/pengembalians": {
            "get": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Ambil semua data pengembalian",
                "operationId": "42183678d877f85754c07865a1ddcd8d",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pengembalian",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "data": [
                                        {
                                            "id": 1,
                                            "peminjaman_id": 1,
                                            "tanggal_dikembalikan": "2026-02-24 13:42:29",
                                            "kondisi_kembali": "baik",
                                            "catatan": "Kembali dalam keadaan utuh",
                                            "peminjaman": {
                                                "id": 1,
                                                "kode": "PINJAM-00001",
                                                "peminjam": {
                                                    "id": 1,
                                                    "name": "User Name"
                                                },
                                                "alat": {
                                                    "id": 1,
                                                    "nama": "Laptop Asus"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Catat pengembalian baru (bisa via Scan QR/Kode atau ID)",
                "operationId": "32dce95d56188ba79bca47d8979f9b0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_peminjaman": {
                                        "description": "Wajib jika peminjaman_id kosong",
                                        "type": "string",
                                        "example": "PINJAM-00001"
                                    },
                                    "peminjaman_id": {
                                        "description": "Wajib jika kode_peminjaman kosong",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "kondisi_kembali": {
                                        "type": "string",
                                        "enum": [
                                            "baik",
                                            "rusak",
                                            "hilang"
                                        ],
                                        "example": "baik"
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Alat kembali lengkap"
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alat berhasil dikembalikan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Alat berhasil dikembalikan.",
                                    "data": {
                                        "id": 1,
                                        "peminjaman_id": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pengembalians/{id}": {
            "get": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Ambil detail pengembalian",
                "operationId": "e9ff8f71fbde44a90b35c3c6ebecfc8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail pengembalian"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Hapus data pengembalian (Membatalkan pengembalian)",
                "operationId": "23714bc20a300d7783ca7d50445dee9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pengembalian berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Update data pengembalian",
                "operationId": "64018b6b52a2751765506e0d115d8a78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kondisi_kembali": {
                                        "type": "string",
                                        "enum": [
                                            "baik",
                                            "rusak",
                                            "hilang"
                                        ]
                                    },
                                    "catatan": {
                                        "type": "string"
                                    },
                                    "tanggal_dikembalikan": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pengembalian berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pengembalians/trashed": {
            "get": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Ambil data pengembalian yang dihapus (Trash)",
                "operationId": "134ee048215ed055312d7688eb9a496f",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data trashed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pengembalians/{id}/restore": {
            "post": {
                "tags": [
                    "Pengembalian"
                ],
                "summary": "Kembalikan data pengembalian yang dihapus",
                "operationId": "5b8cd00dd191dd54690f02c7fa4c2c53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil di-restore"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}